# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AcidResponse:
    properties:
      hplus:
        type: integer
      id:
        type: integer
      img:
        type: string
      info:
        type: string
      molar_mass:
        type: number
      name:
        type: string
      name_ext:
        type: string
    type: object
...
  dto.CarbonateAcidResponse:
    properties:
      acid:
        $ref: '#/definitions/dto.AcidResponse'
      acid_id:
        type: integer
      carbonate_id:
        type: integer
      id:
        type: integer
      mass:
        type: number
      result:
        type: number
    type: object
  dto.CarbonateDetailResponse:
    properties:
      acids:
        items:
          $ref: '#/definitions/dto.CarbonateAcidResponse'
        type: array
      creator:
        type: string
      date_create:
        type: string
      date_finish:
        type: string
      date_update:
        type: string
      id:
        type: integer
      mass:
        type: number
      moderator:
        type: string
      status:
        type: string
    type: object
  dto.CarbonateIconsResponse:
    properties:
      acid_count:
        type: integer
      carbonate_id:
        type: integer
    type: object
  dto.CarbonateListResponse:
    properties:
      carbonates:
        items:
          $ref: '#/definitions/dto.CarbonateResponse'
        type: array
    type: object
  dto.CarbonateResponse:
    properties:
      creator:
        type: string
      date_create:
        type: string
      date_finish:
        type: string
      date_update:
        type: string
      id:
        type: integer
      mass:
        type: number
      moderator:
        type: string
      status:
        type: string
    type: object
  dto.CarbonateUpdateRequest:
    properties:
      mass:
        minimum: 0
        type: number
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: kovalyovea@student.bmstu.ru
    name: API Support
    url: https://github.com/nylox23/DIA
  description: Calculates the volume of gas emitted in a reaction
  license:
    name: AS IS (NO WARRANTY)
  title: Carb_Calc
  version: "1.0"
paths:
...
  /api/carbonates:
    get:
      description: Возвращает список заявок, отфильтрованный по статусу и дате
      security:
        - Bearer: []
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: Фильтр даты от
        in: query
        name: date_from
        type: string
      - description: Фильтр даты до
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CarbonateListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список заявок
      tags:
      - carbonates
    put:
      consumes:
      - application/json
      description: Обновляет поле "масса карбоната" в текущей заявке-черновика пользователя
      security:
        - Bearer: []
      parameters:
      - description: Обновляемые данные заявки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CarbonateUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить заявку
      tags:
      - carbonates
  /api/carbonates/{id}:
    delete:
      description: Удаляет заявку (создателем - только черновики или отклоненные заявки)
      security:
        - Bearer: []
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить заявку
      tags:
      - carbonates
    get:
      description: Возвращает заявку со списком добавленных в нее кислот
      security:
        - Bearer: []
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CarbonateDetailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить заявку по ID
      tags:
      - carbonates
  /api/carbonates/{id}/status:
    put:
      consumes:
      - application/json
      description: Устанавливает статус заявки на "завершен" или "отклонен"
      security:
        - Bearer: []
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Установить статус заявки
      tags:
      - carbonates
  /api/carbonates/current:
    get:
      description: Возвращает данные по текущей заявке-черновику пользователя для
        кнопки перехода
      security:
        - Bearer: []
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CarbonateIconsResponse'
      summary: Получить данные текущей заявки
      tags:
      - carbonates
  /api/carbonates/form:
    put:
      description: Меняет статус текущей заявки-черновика пользователя на "сформирован"
      security:
        - Bearer: []
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Сформировать заявку
      tags:
      - carbonates
...
schemes:
- http
swagger: "2.0"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header